Adds pkg/cache, a type-safe way to store GCP objects
This commit adds the data structure, follow on commits will use the data
structure to implement both a cache of the cloud state as well as use it
to implement the synchronization routines. See the package docstring in
cache.go for more details.

The cache code is type-safe and generated from descriptions contains in
the pkg/cache/meta. The contents of `gen.go` are derived from `$ go run
gen/main.go > gen.go`

As the code is not expected to change often (only when types are
added/removed), the generated code is committed to the repo and the code
generation is expected to be run by hand. If there are significant
divergences from these assumptions, then full integration into the build
process will be needed.