Update deploy/glbc yaml files for BackendConfig
Not sure if BackendConfig support was intentionaly left out of the deployment
example (ie. after all it's still off by default, that feature is possibly not
yet considered ready for usage outside GKE?), or this example was just forgotten.

If the later, including it would be nice as those GKE users having already
defined BackendConfigs and willing to try ingress-gce git HEAD would be up
for a suprise.

Otherwise, let's just drop this PR.