Added --dry-run
This is very useful for testing. It uses kubectl --dry-run where
possible, so we're actually testing the structure of the YAML files
(the creation of the key does not use this since it depends on
running a real command).

This command can also be used to quickly and cleanly get the commands
this script would run for manual usage.

Fixed bug with --cleanup  where it would ignore any args after the
--cleanup tag (which could result in running the wrong things!).