Include EndpointsCalculatorMode in negSyncerKey.
This is neeeded to identify a service TrafficPolicy change as an
event to stop the old syncer and start a new one with the new mode.
This is relevant only for VM_IP NEGs. The field will have the value
"L7" for the VM_IP_PORT NEGs.

Changed portInfoMap as well.