Add a healthcheck for L4 ILB controller.
Since there is no guaranteed ILB service that the controller will always sync, we cannot
check for an absolute lastSync timestamp. Instead, this change tracks the last enqueue
timestamp and ensures that the last sync happened soon(15 mins later atmost) after enqueue.
This change only logs an error and does not fail the healthcheck in case of the delay. Returning error will be addressed in a follow-up.