Enqueue node updates to VM_IP_NEG controller.
This is needed in order to include nodes that are unschedulable/unready
at creation time, but change to a ready status later. If this update is
missed, the NEGs in Cluster traffic policy will not include all eligible
nodes. This is especially an issue if nodes are added in a new zone
where no NEG exists currently.