Prevent race conditions between service and subsetting controllers.
If the service finalizer fields are manually reset, the subsetting controller
will add the v2 finalizer, thinking it is a new service. The service controller
will also add the v1 finalizer because the legacy forwarding rule still exists.
This results in unnecessary creation of NEGs and other LB resources for the legacy
ILB. This change prevents this race.