Skip unsupported healths during NEG polling
Currently there are 4 types of healths a network endpoint can have:
  * forwardingRule
  * backendService
  * healthCheck
  * healthCheckService

Currently, the NEG controller explicitly checks for a backendService,
however if there is any of the other healths, rather than fast-failing
it will timeout waiting for a backendService to appear and become
healthy.

This is really just a workaround, and ideally the code would be
refactored to support all possible healths. For now, this will bring
back correctness to the fast-fail waiting on backendServices iff
backendServices exist in healths.