Fix fo firewall rule deletion
When L4 LoadBalancer is deleted firewall rule was deleted upfront.
But when ExternalTrafficPolicy is set to Cluster health check is shared
among all ILB or XLB services and removing firewall rule will make those
services unreachable. So in service deletion we need to check if health
check is not in use by another service,