Edit L4 Namer to support External and Internal Load Balancer
Function creating names for ILB aligning with k8s2 convention is
VMIPNEG.
Since we use the same name for neg, backend service and firewall rules I
rename this function to L4Backend. Now we can use the same function for
NetLB services and NetLB does not use NEG so VMIPNEG name is not
suitable here.