Add new flag for configuring metrics export frequency
We need to increase the frequency to at least 1/5minutes, but this may not be the final value.
From observability POV the more frequent the better.
With this change we can tune the frequency without rebuilding the binary, and without releasing new tags.

This will allow us to increase export frequency to comply with sampling theorem
https://en.wikipedia.org/wiki/Nyquist%E2%80%93Shannon_sampling_theorem