ConfigMap informer should only watch --asm-configmap-based-config-namespace
This avoids making total ConfigMap size part of the scalability
profile of the controller.

Testing: Manual

- hack/run-local-glbc.sh --enable-asm-config-map-config \
-v=3 --asm-configmap-based-config-namespace=default
- Created ConfigMaps in `default` namespace.
- Observed only updates to the default namespace ConfigMaps were being
observed in the controller.

Notes

One interesting thing to note is that watching kube-system has a
non-trivial amount of constant activity due to leader election leases.
This can cause CPU usage if there are a large number of components
running leader election using the same namespace.