[Cherrypick #1828] l4 healthcheck timeouts change
This change adds changed l4 healthcheck timeouts
to the v1.14 release.
According to l4 healthcheck constants
the default set up has timeout = 8 seconds
and threshold = 3 times. As result it can
lead to 24 (8*3) seconds of lost requests.
This change is supposed to decrease this
time to only 6 seconds for sharedHc=false.

Signed-off-by: Elina Akhmanova <elinka@google.com>