Add check for endpoint counts
Add check to make sure the total number of endpoints from endpointData/EPS
equals to the number of endpoints from endpointPodMap. If not, the
syncer will enter the error state.
EndpointData may contain duplicated endpoint, but endpointPodMap removes
duplicated endpoints, and dupCount stores the number of duplicated it removed.
We compare the endpoint counts with duplicates.
Change the function definition of NetworkEndpointsCalculator.CalculateEndpoints()
to return another variable that represents the count of duplicated endpoints.