Clean up error state checking
Consolidate EndpointSlice nodeName and zone check, and flip checking
logic for better readability.
Return error if nodeName or zone is missing when CalculateEndpoints, and
check for error state based on the returned error. Add check for case
when nodeName is set to be empty string. Flip invalid check to valid
check.