Trim the path after status code
Currently GKE Gateway controller gxlb use urlRewrite to rewrite pathPrefix to "/statuscode/500" and set backend service to this 404 server.

For example, an HTTPRoute specifies path prefix match of "/" with an invalid backend reference. A request to path "/foo" would get rewritten to "/statuscode/500/foo".

However, Ingress 404 server currently uses exact match in statusCodeHandler() to generate the status code to be returned. When request URL path is "/statuscode/500", 500 is returned. But with path being "/statuscode/500/foo", an error occurs and 404 is returned instead.

We need to update 404 server to treat "/statuscode/500" as a prefix.