l4 controller deletion vs mutation
The l4 controller can stop processing a Service in two cases, when it
has been deleted (it has to remove the finalizer) or it has being
mutated and is no longer managed by the controller.

If the Service is being mutated the controller removes the custom
annotations and status before removing the finalizer.