NEG Controller - run syncers per network
This commit adds the service network to the syncer key.
The goal is to make the NEG controller differentiate between service ports in different networks. Currently, if the service network changes the NEG controller doesn't know the ports are really different and keeps the old syncers and so is unable to create appropriate NEGs.
With this change it properly recognizes the change, cleans up the invalid NEG and creates a new one in the proper network.

When multinet is turned off, the network in the key will always be `default` and it will not affect the logic.