IG Manager: deduplicate IG names when listing
The manager sync function assumes the names of the groups are listed without a zone and will attempt to do the add/remove operations in all zones.
Before this change, if the group existed in more than one zone it would be returned many times (list will contain the same name a few times). This leads to unnecessary iterations and instances listing - everything would happen N^2 for N=number of zones.