Do not consider forbidden errors within the PSC controllers sync metrics
Previously, we used to see less of 403 forbidden errors because we
always used to look for subnets in the same project as the GKE clusters
project, in which case the service-account always had permissions. This
meant that for Shared VPC use cases, we used to incorrectly look for the
subnet in the service project and not in the host project.

After the changes introduced in commit
https://github.com/kubernetes/ingress-gce/commit/12ee6f29c07f6542a49e4abd82a36a8046999612#diff-aacf8e53dbfd73dd635a4d6615f52725b9f2b1ebcdbb25efae5a6bae20201fd7 within the file vendor/k8s.io/cloud-provider-gcp/providers/gce/support.go, we now corrected the behaviour for Shared VPC where we now start looking for the subnet in the host project. But this often leads to 403 forbidden errors because it requires the user to have explicitly given the GKE service-account permissions on the Shared VPC host project, which the customers often forget to do.