Modify the regional IG linker to not detach instance gropus from zones that still have nodes.
There is an issue where after master upgrade all nodes are seen as unready by the master. This causes the controller to see this state, these nodes seem to be unfit candidates for LB backends and the controller removes the matching instance groups. Theis causes traffic outage in the LB.

This change prevents removing instance groups for zones where some nodes exist, even if these nodes are unready.