Use all zones when creating the shared IG for a service.
With this change, the code to ensure IGs will use AllNodesFilter to determine which zones to create IGs in.
The goal is to avoid detaching groups which become temporarily unready, for example after master upgrade.
This change should be fine since creating and using an IG that could be empty is ok from the LB perspective. On the other hand removing a temporarily unready group can cause 10 minutes of no traffic.