Detect GCE operations that are unexpected for the Subsetting ILB.
Code to detect a situation when a service processing is a resync but causes GCE resource
updates.
This also touches the NetLB part, especially for the dual stack case, where some resources can be doubled (for example forwarding rule). Modification of just one should mark the resource type as updated.