Fix port range to port comparison for forwarding rules
This fixes a bug when an existing forwarding rule lists a distinct a set of ports.
If a new port is added to a load balancer that falls within the same range of ports,
the forwarding rule will not be updated because we would compare the old and new list
of ports as a port range.

For example, with oldPorts = {1, 5, 10} and newPorts = {1, 5, 6, 10}, both old and new
ports are converted to port range 1-10 and no change is made to the forwarding rule.

With this change we only convert new ports to a port range if the existing ports are
also a port range.