Fix forwarding rule resources leak when switching protocol and stack
This issue is affecting switching between both stack and protocol at the same time. For example:
* IPv4 TCP -> IPv6 UDP,
* IPv4 UDP -> IPv6 L3,
* IPv6 TCP -> IPv4 UDP,
* Dual Stack TCP -> IPv4 UDP.
This is because it tries to generate old forwarding rule name based on the new protocol. For example for IPv4 TCP -> IPv6 UDP it will try to delete IPv4 UDP (which doesn't exist) forwarding rule, not IPv4 TCP.

This is fixed by trying to delete all potential forwarding rules for ILB. 404 not found errors are ignored by firewall provider struct.