Partially delete NEGs and update NEG CR
When a subnet is removed, the relevant NEGs will be marked with the
TO_BE_DELETED state. NEGs need to be deleted, before the relevant subnet
can be removed from the cluster or deleted. Therfore, during GC, we
should attempt deletion of these specific NEGs, and should not delete
any NEGs that are in the ACTIVE or INACTIVE state.

Improve logging and error messages during garbage collection