fix: NEG Readiness gate for MSC is no longer prematurely marked ready
Previously, when we didn't create NEGs for the additinoal subnets, we had the
logic in place to make sure that the pods which exist in nodes from the
additional subnet are always marked ready for the NEG readiness gate. This was
done since the pod will never get added to a NEG and hence won't have it's
readiness gate ready in a natural way.

But now that we DO have NEGs in additional subnets, we no longer need this
explicit handling.

This change is essentially a manual revert of the commit
https://github.com/kubernetes/ingress-gce/pull/2560/commits/f9895f1beb92fc81ad81915565e7dbc7e3add38c