Create legacy mode for zone getter for legacy network clusters
In Legacy mode, subnets do not exist and MSC is naturally incompatible
with legacy network based GKE clusters. Therefore when in legacy mode
is detected (when the provided SubnetworkURL is empty), then remove
all validation related to subnetworks. This includes checks on the
PodCIDR. We assume that all nodes exist in the legacy network and the
only checks done are on the ProviderID and conditionally based on
the readiness.