Add Compare for sets of Forwarding Rules
This allows to compare two slices of Forwarding Rules. It has Equal and Patchable as parameters to allow it to be used for both IPv4 and IPv6 which have different func implementations.

We also add tests for verifying logic specific to compare, as exact equal/patchable logic are tested elsewhere.