This commit introduces three new folders inside the cmd directory.
The purpose is to create three new binaries that willl be subsituting the monolithic binary that existed until today (e2e-test).
The new binaries are ingress-controller-e2e-test, neg-e2e-test, and psc-e2e-test.

In particular each new folder hosts: the code of the specific tests for the respective controller; an entry point for the container/binary; an upgrade folder with the targets for the upgrade_test.go.

In addition three new Docker templates are introduced to build the three images.

Finally Makefile is updated with the new binaries.

The old e2e is preserveed and it will be deleted later once the new images are tested and proven to work correctly within the CI/CD pipeline.