Fix IAP OAuth credentials not being fetched when IAP is disabled
The behavior in validateIAP() of treating .spec.iap.enabled=false as
equivalent to .spec.iap=nil and abort all validation is incongruent with
that in pkg/backends/features/iap.go, where a full sync of the IAP
configuration is triggered as long as .spec.iap!=nil.

The latter behavior is the correct one, as it allows disabling IAP on an
ingress where it was previously enabled by setting
.spec.spec.enabled=false in the BackendConfig.

However, because pkg/backendconfig/validation.go skips the entire
validation and with that the lookup of the OAuth credentials referenced
in .spec.iap.oauthclientCredentials.secretName, the desired
BackendConfig's OAuth Client ID will be uninitialized and interpreted
as empty (""). And in the case that IAP was previously configured with a
non-default Client ID, this will trigger a switchingToDefaultError.