Gracefully handle empty default backend service / gate NEGs for Ingress
This commit prevents controllers initialization from crashing and controllers performing unnecessary work by ensuring the DefaultBackendServicePort function returns an empty utils.ServicePort when the Ingress controller is disabled. This change modifies the core behavior of the DefaultBackendServicePort function, making it resilient to cases where the default backend service does not exist when the Ingress controller is not running.

It also ensures that the NEG controller doesn't create NEGs for Ingress when the GKE Ingress controller is disabled together with ensuring that the controller's informers for Ingress-related objects remain inactive, eliminating unnecessary resource consumption.