Implement a legacy mode for the zonegetter
Legacy networks do not have the concept of subnets and naturally are not
compatible with the multi subnet feature. Therefore zonegetter must be
able to handle legacy networks and not error out. Legacy mode zonegetter
still validates that the Nodes have the pod cidr and the provider ID set
before using it for any calculation

Legacy Network Assumptions:
* All Nodes are considered part of the default network
* The subnetwork is always empty

Adds a readme on what the zonegetter does at a high level