Reduce the number of Endpoints for ILB ETP Cluster
24 is a better number than 25 as its divisible by 1, 2, 3, 4 and 6. This should lead to reduced detach - attach of endpoints in situations when we have two Zones with similar number of Nodes. In previous implementation the Zone with the most amount of Nodes would receive an extra endpoint. If other zone had suddenly increased the number of its Nodes to beat the previous leader it would cause a detach of an endpoint from previous leader with 30 second connection draining, followed by attach to the new leader. If Nodes fluctuate frequently this would cause a lot of unnecessary connection draining.

To avoid that this commit reduces the amount of Endpoints that are used to 24 for clusters that operate on 1, 2, 3, 4 or 6 zones. For clusters running on other number of Zones we use 25 endpoints.

Additionally it makes the removal of endpoints deterministic by sorting endpoints before they are used. It wasn't previously deterministic as there was a conversion from map, which doesn't have an order.

(cherry picked from commit b4518094c509697a7aa57ed7814aecce462a6f56)